More code cleanup
