+++ libflist/FChat/Commands/Server/Character.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using libflist.Connection.Types;
-using libflist.Util.Converters;
+﻿using libflist.Util.Converters;
 using Newtonsoft.Json;
 
 namespace libflist.FChat.Commands.Server.Character
