--- libflist/FChat/Connection/Types/Channel.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace libflist.Connection.Types
-{
-	public enum ChannelMode
-	{
-		Chat,
-		Ads,
-		Both
-	}
-
-	public enum ChannelStatus
-	{
-		Public,
-		Private
-	}
-}
