--- libflist/FChat/Connection/Types/ResponseType.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace libflist.Connection.Types
-{
-	public enum ResponseType
-	{
-		Default,
-		None,
-		Multiple
-	}
-}
-
