--- libflist/FChat/Connection/Types/UserRight.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace libflist.Connection.Types
-{
-	public enum UserRight
-	{
-		Disconnected = -1,
-
-		User = 0,
-		ChatOP,
-		Admin
-	}
-}
-
